+++ cmd/experiments/2026-02-08--simulated-communication/plugin-playground/packages/plugin-runtime/src/index.ts
@@ -1,0 +1,6 @@
+export * from "./contracts";
+export * from "./dispatchIntent";
+export * from "./runtimeIdentity";
+export * from "./runtimeService";
+export * from "./uiSchema";
+export * from "./uiTypes";